+++ BitTorrent/Tracker/TorrentUDPTracker.swift
@@ -40,7 +40,7 @@ class TorrentUDPTracker: TorrentTracker {
     }
     
     func announceClient(with peerId: String,
-                        port: Int,
+                        port: UInt16,
                         event: TorrentTrackerEvent = .started,
                         infoHash: Data,
                         numberOfBytesRemaining: Int,
